Improve theme handling
