+++ ui/App.svelte
@@ -9,6 +9,7 @@
   import { NotificationFaucet, Remote } from "./DesignSystem/Component";
 
   import Hotkeys from "./Hotkeys.svelte";
+  import Theme from "./Theme.svelte";
 
   import Blank from "./Screen/Blank.svelte";
   import IdentityCreation from "./Screen/IdentityCreation.svelte";
@@ -73,6 +74,7 @@
 
 <Hotkeys />
 <NotificationFaucet style="margin-top: calc(var(--topbar-height) + 11px)" />
-<Remote {store} context="session">
+<Theme />
+<Remote {store} let:data={session} context="session">
   <Router {routes} />
 </Remote>
