+++ ui/Hotkeys.svelte
@@ -2,8 +2,6 @@
   import { location, pop, push } from "svelte-spa-router";
 
   import * as path from "./src/path.ts";
-  import * as session from "./src/session.ts";
-  import * as theme from "./src/theme.ts";
 
   const toggle = (destination) => {
     if (path.active(destination, $location)) {
@@ -15,10 +13,6 @@
   const onKeydown = (event) => {
     if (event.target !== document.body) {
       return false;
-    }
-
-    if (event.shiftKey && event.code === "KeyC") {
-      session.updateAppearance({ theme: theme.next() });
     }
 
     if (event.shiftKey && event.code === "KeyD") {
