+++ ui/src/settings.ts
@@ -1,7 +1,11 @@
-import * as theme from "./theme";
+// TYPES
+export enum Theme {
+  Dark = "dark",
+  Light = "light",
+}
 
 export interface Appearance {
-  theme: theme.Theme;
+  theme: Theme;
 }
 
 export enum Network {
@@ -18,3 +22,34 @@ export interface Settings {
   appearance: Appearance;
   registry: Registry;
 }
+
+interface Option {
+  value: string;
+  title: string;
+}
+
+export  const networkOptions: Option[] = [
+  {
+    value: Network.Emulator,
+    title: "Emulator",
+  },
+  {
+    value: Network.FFnet,
+    title: "FFnet",
+  },
+  {
+    value: Network.Testnet,
+    title: "Testnet",
+  },
+];
+
+export const themeOptions: Option[] = [
+  {
+    title: "Light",
+    value: Theme.Light,
+  },
+  {
+    title: "Dark",
+    value: Theme.Dark,
+  },
+];
